Make the run config in task_output a properly typed object. Will make our lives a lot better going forward
